feat/NPCModule (#266)
* +


* use jpenillas resource factory so the manual yml is unneeded.


* stdlib instead of stdlib-jdk8


* tweaking


* fix version and using JsonMapper instead


* making configmanager use datainterface and refactoring datainterface to be more generic


* woops


* added module.key() util


* [ci-skip] Update gradle/actions digest to 6a96db5 (#259)


* [ci-skip] Update plugin xyz.jpenilla.run-paper to v3.0.2 (#260)


* fmt


* [ci-skip] Update softprops/action-gh-release digest to 6da8fa9 (#261)


* playerskull not displayname but name instead


* init


* +


* +


* +

* +

* edit var name


* +


* some check


* +


* fix config


* added head before name in chat


* rename


* fix sprite


* fix order


* fmt


* expressions and handling in config


* fmt


* refactor


* added sounds to the trade


---------

+++ src/main/kotlin/org/xodium/vanillaplus/modules/PetModule.kt
@@ -81,8 +81,11 @@ internal class PetModule : ModuleInterface<PetModule.Config> {
      */
     private fun findLeashedPet(player: Player): Tameable? =
         player
-            .getNearbyEntities(config.transferRadius, config.transferRadius, config.transferRadius)
-            .filterIsInstance<LivingEntity>()
+            .getNearbyEntities(
+                config.transferRadius.toDouble(),
+                config.transferRadius.toDouble(),
+                config.transferRadius.toDouble(),
+            ).filterIsInstance<LivingEntity>()
             .firstOrNull { it.isLeashed && it.leashHolder == player }
             as? Tameable
 
@@ -114,7 +117,7 @@ internal class PetModule : ModuleInterface<PetModule.Config> {
 
     data class Config(
         override var enabled: Boolean = true,
-        var transferRadius: Double = 10.0,
+        var transferRadius: Int = 10,
         var i18n: I18n = I18n(),
     ) : ModuleInterface.Config {
         data class I18n(
